Complete

+++ main/models.py
@@ -6,4 +6,4 @@ class User(models.Model):
     '''Creates a table with username, room, and score'''
     username = models.CharField(max_length=50)
     room = models.CharField(max_length=50)
-    score = models.IntegerField(default=0)
+    score = models.DecimalField(default=0.0, decimal_places=3, max_digits=100)
